All scalar values are floating point numbers by default

+++ include/afml/nn/Activations/Threshold.hpp
@@ -17,7 +17,7 @@ namespace afml
         class ThresholdNode : public ActivationNode
         {
         private:
-            double mVal;
+            float mVal;
 
             af::array fn(const af::array &input)
             {
@@ -30,7 +30,7 @@ namespace afml
                 return (input >= mVal).as(input.type());
             }
         public:
-            ThresholdNode(int size, double val, const char *name="none") :
+            ThresholdNode(int size, float val, const char *name="none") :
                 ActivationNode(size, name),
                 mVal(val)
             {
